requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
groq>=0.9.0
hindsight-client>=0.3.0
python-dotenv>=1.0.0
pdfplumber>=0.11.0
